requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
Faker==20.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
